[dict-proto] C++ implementation of SwissNameDictionary, pt. 6
This is a reland of
https://chromium-review.googlesource.com/c/v8/v8/+/2715192.

This CL is part of a series that adds the C++ implementation of
SwissNameDictionary, a deterministic property backing store based on
Swiss Tables.

This CL adds the Add, DeleteEntry, Rehash, and Shrink operations, as
well as some helpers required by those.

Bug: v8:11388
Change-Id: I028dbf3215c76e4f40cba420c125ef25223f252e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2719561
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Frank Emrich <emrich@google.com>
Cr-Commit-Position: refs/heads/master@{#73040}